remove conversion warning
git-svn-id: https://svn.mech.kuleuven.be/repos/orocos/orocos-trunk@2436 ce417995-dfc9-0310-95a0-acdaff106893

+++ tests/tasks_multiple_test.cpp
@@ -74,7 +74,7 @@ CPPUNIT_TEST_SUITE_REGISTRATION( TasksMultipleTest );
 
         // we lower the 'load' if the period is too short.
         // this is a bit arbitrary.
-        int correction = 0.001 / ( ZeroTimeThread::Instance()->getPeriod() );
+        int correction = int(0.001 / ( ZeroTimeThread::Instance()->getPeriod() ));
         if ( correction == 0)
             correction = 1;
         if ( correction > 20 )
